# docs/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "REST API with authentication, roles and encryption.",
        "title": "Go API Dojo",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8000",
    "basePath": "/",
    "paths": {
...
        "/api/v1/chapters": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "chapters"
                ],
                "summary": "Create a new chapter",
                "parameters": [
                    {
                        "description": "Chapter payload",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.CreateChapterRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/model.CourseChapter"
                        }
                    },
                    "400": {
                        "description": "invalid request body",
                        "schema": {
                            "$ref": "#/definitions/fault.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "missing or invalid token",
                        "schema": {
                            "$ref": "#/definitions/fault.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/api/v1/chapters/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "chapters"
                ],
                "summary": "Get a chapter by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Chapter ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.CourseChapter"
                        }
                    },
                    "400": {
                        "description": "invalid id",
                        "schema": {
                            "$ref": "#/definitions/fault.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "chapter not found",
                        "schema": {
                            "$ref": "#/definitions/fault.ErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "chapters"
                ],
                "summary": "Update a chapter",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Chapter ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Chapter payload",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.UpdateChapterRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.CourseChapter"
                        }
                    },
                    "400": {
                        "description": "invalid request body",
                        "schema": {
                            "$ref": "#/definitions/fault.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "chapter not found",
                        "schema": {
                            "$ref": "#/definitions/fault.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "chapters"
                ],
                "summary": "Delete a chapter",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Chapter ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "401": {
                        "description": "missing or invalid token",
                        "schema": {
                            "$ref": "#/definitions/fault.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "chapter not found",
                        "schema": {
                            "$ref": "#/definitions/fault.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/api/v1/courses/{course_id}/chapters": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "chapters"
                ],
                "summary": "List chapters for a course",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Course ID",
                        "name": "course_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.CourseChapter"
                            }
                        }
                    },
                    "400": {
                        "description": "invalid course id",
                        "schema": {
                            "$ref": "#/definitions/fault.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/fault.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "fault.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "a descriptive error message"
                }
            }
        },
...
        "handler.CreateChapterRequest": {
            "type": "object",
            "required": [
                "course_id",
                "slug",
                "title"
            ],
            "properties": {
                "course_id": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "duration_minutes": {
                    "type": "integer",
                    "minimum": 0
                },
                "is_free": {
                    "type": "boolean"
                },
                "is_published": {
                    "type": "boolean"
                },
                "slug": {
                    "type": "string",
                    "maxLength": 255,
                    "minLength": 2
                },
                "sort_order": {
                    "type": "integer"
                },
                "title": {
                    "type": "string",
                    "maxLength": 255,
                    "minLength": 2
                }
            }
        },
...
        "handler.UpdateChapterRequest": {
            "type": "object",
            "required": [
                "slug",
                "title"
            ],
            "properties": {
                "description": {
                    "type": "string"
                },
                "duration_minutes": {
                    "type": "integer",
                    "minimum": 0
                },
                "is_free": {
                    "type": "boolean"
                },
                "is_published": {
                    "type": "boolean"
                },
                "slug": {
                    "type": "string",
                    "maxLength": 255,
                    "minLength": 2
                },
                "sort_order": {
                    "type": "integer"
                },
                "title": {
                    "type": "string",
                    "maxLength": 255,
                    "minLength": 2
                }
            }
        },
...
        "model.CourseChapter": {
            "type": "object",
            "properties": {
                "courseID": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "durationMinutes": {
                    "type": "integer"
                },
                "id": {
                    "type": "string"
                },
                "isFree": {
                    "type": "boolean"
                },
                "isPublished": {
                    "type": "boolean"
                },
                "slug": {
                    "type": "string"
                },
                "sortOrder": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Format: Bearer {token}",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}